move yaml gen to cd
